extract mutators
